feat: add if for copyFolder

+++ 04-copy-directory/index.js
@@ -2,7 +2,7 @@ const fs = require("fs");
 const { copyFile } = require("fs/promises");
 const path = require("path");
 
-function copyDir() {
+async function copyDir() {
   const filesFolder = path.join(__dirname, "/files");
   const copyFolder = path.join(__dirname, "/files-copy");
 
@@ -12,15 +12,15 @@ function copyDir() {
   });
 
   fs.readdir(copyFolder, { withFileTypes: true }, (err, files) => {
-    // if (files.length !== 0) {
-    files.forEach((file) => {
-      if (file.isFile()) {
-        fs.unlink(path.join(copyFolder, path.basename(file.name)), (err) => {
-          if (err) throw err;
-        });
-      }
-    });
-    // }
+    if (files.length !== 0) {
+      files.forEach((file) => {
+        if (file.isFile()) {
+          fs.unlink(path.join(copyFolder, path.basename(file.name)), (err) => {
+            if (err) throw err;
+          });
+        }
+      });
+    }
   });
 
   fs.readdir(filesFolder, { withFileTypes: true }, (err, files) => {
